Fixed most issues, moved filtering to MetaMorpheus Task
