+++ MetaMorpheus/Test/SpectralRecoveryTest.cs
@@ -133,7 +133,6 @@ namespace Test
         [Test]
         public static void SpectralRecoveryPostSearchAnalysisTest()
         {
-
             List<string> warnings;
             string mbrAnalysisPath = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestSpectralRecoveryOutput\SpectralRecovery\RecoveredSpectra.psmtsv");
             string expectedHitsPath = Path.Combine(TestContext.CurrentContext.TestDirectory, "TestData", @"SpectralRecoveryTest\ExpectedMBRHits.psmtsv");
